+++ __tests__/__integration__/cmd/src/cli/validate/syntax/Syntax.definition.ts
@@ -52,6 +52,12 @@ export const syntaxTestCommand: ICommandDefinition = {
                 allowableValues: {values: ["allowableA", "allowableB", "^allowableC\\$"], caseSensitive: false}
             },
             {
+                name: "option-to-specify-4",
+                description: "Part of must specify one group",
+                type: "array",
+                allowableValues: {values: ["allowableA", "allowableB"], caseSensitive: false}
+            },
+            {
                 name: "conflicts-with-1",
                 description: "Conflicts with option-to-specify-2",
                 type: "string",
@@ -161,5 +167,5 @@ export const syntaxTestCommand: ICommandDefinition = {
                 options: `"file.txt" "ibmuser.ps" --mr wait`
             },
         ],
-    mustSpecifyOne: ["option-to-specify-1", "option-to-specify-2", "option-to-specify-3"]
+    mustSpecifyOne: ["option-to-specify-1", "option-to-specify-2", "option-to-specify-3", "option-to-specify-4"]
 };